TM-7670 Tweaks to the Application Restart I made the following changes:   - The POST was going to the wrong action thererfore there was a 404 error   - Fixed up the wording and styling a bit   - The redirect was going to /, added a named URL in URLMappings and redirect user back to Admin Portal   - Added an Audit Builder method (not working at the moment)
The Javascript logic does not properly handle the response and always assumes that it has been successful. It still needs work to check for the status return code and should display the error message instead of saying that the application is restarting.  The POST should post as JSON so that the error handler for 404 isn't the HTML page but just the straight 404.