TM-9590 Fix : Change the DataScript designer so that the Test button is remains active after clicking
There was a rule inside of <check-action> component to always disable it if the state is valid, some times this rule is not desirable by the host component (TM-9509).
Additionally the responsibility to disable <check-action>  was shared by the component itself (disabling if it has a valid state)  and its host component (passing a [disabled] @Input).
The change consisted on remove this duplicated responsibility from <check-action>, and let the host which  has the model to decide if  <check-action> should be disabled through the already  used @Input [disabled] property.