TM-15118 upgrading to latest point release of Grails 3: https://github.com/grails/grails-core/releases/tag/v3.3.10
Found a minor issue with the index.gsp taking over for the tdstm/ route
it's just the default file and removing it fixed the issue.

I also found an issue in the main template, where it was referencing a
permission that didn't exist. Not sure why this didn't cause and error
before.