TM-16282 This is also validating the asset/dependent causing the error
in one of the dependents there is some bad data that probably got there,
from a previously fixed issue. turning off deep validate makes it so the
error doesn't pop up. However if you try to edit and save the current
dependent asset record it will fail with the same error, but it you change
it in the ui using the date picker you'll get the right format and it will
work.