TM-16500 Refactoring to remove the duplicated function filterColumn
By the moment filterColumn of providers services was removed and
we point out to filterColumn of data-list-grid.model.
In the future the function that will prevail handling all filtering logic
is the one defined at DataGridOperations helper, having more time to do
a bigger refactoring over the grids that are not  using DataGridOperations Helper.